[OBMIN-397] added tooltip to components
