+++ src/components/ellipsis-text/arg-types.ts
@@ -8,15 +8,7 @@ export const argTypes: IArgTypes<ITooltipProps> = {
     type: { name: 'string' },
     description: 'Tooltip reference element',
   },
-  id: {
-    name: 'Id',
-    type: { name: 'string' },
-    description: 'This prop is used to help implement the accessibility logic.',
-    table: {
-      type: { summary: 'string' },
-    },
-  },
-  place: {
+  position: {
     name: 'place',
     type: { name: 'string' },
     description: 'Tooltip placement.',
@@ -27,42 +19,12 @@ export const argTypes: IArgTypes<ITooltipProps> = {
     control: { type: 'select' },
     options: ['top', 'bottom', 'left', 'right'],
   },
-  offset: {
-    name: 'Offset',
-    type: { name: 'number' },
-    description: `Space between the tooltip element and anchor element`,
-    table: {
-      defaultValue: { summary: 10 },
-      type: { summary: 'number' },
-    },
-    control: { type: 'number' },
-  },
-  width: {
-    name: 'width',
+  widthTooltip: {
+    name: 'widthTooltip',
     type: { name: 'number' },
     description: 'Tooltip CSS width ',
     table: {
       defaultValue: { summary: 200 },
-      type: { summary: 'number' },
-    },
-    control: { type: 'number' },
-  },
-  delayShow: {
-    name: 'delayShow',
-    type: { name: 'number' },
-    description: `CSS The number of milliseconds to wait before showing the tooltip. This prop won't impact the enter touch delay`,
-    table: {
-      defaultValue: { summary: 500 },
-      type: { summary: 'number' },
-    },
-    control: { type: 'number' },
-  },
-  delayHide: {
-    name: 'delayHide',
-    type: { name: 'number' },
-    description: `The number of milliseconds to wait before hiding the tooltip. This prop won't impact the leave touch delay`,
-    table: {
-      defaultValue: { summary: 500 },
       type: { summary: 'number' },
     },
     control: { type: 'number' },
