+++ src/components/index.tsx
@@ -4,8 +4,7 @@ export { Button } from './button';
 export { CancelEverything } from './cancel-everything';
 export { CheckBox } from './checkbox';
 export { Deals } from './deals';
-// TODO: remake ellipsis-text
-// export { EllipsisText } from './ellipsis-text';
+export { EllipsisText } from './ellipsis-text';
 export { HelpChildren } from './help-children';
 export { Image } from './image';
 export { InboxMessage } from './inbox-message';
@@ -25,7 +24,6 @@ export { ProductPostData } from './product-post-data';
 export { Subtitle } from './subtitle';
 export { Title } from './title';
 export { Toast, showMessage } from './toast';
-export { Tooltip } from './tooltip';
 export * as Icon from './icon';
 export * as Images from './img';
 export * as Responsive from './responsive';
