+++ src/components/product-owner-data/index.tsx
@@ -1,7 +1,6 @@
 import * as Styles from './styles';
 import { Avatar } from '../avatar';
-// TODO: remake ellipsis-text
-// import { EllipsisText } from '../ellipsis-text';
+import { EllipsisText } from '../ellipsis-text';
 import { IProductOwnerDataProps } from './types';
 
 const ProductOwnerData = ({
@@ -23,9 +22,7 @@ const ProductOwnerData = ({
 
         <Styles.ProductOwnerName>
           <Styles.ProductOwnerTitleH2>
-            {/* <EllipsisText id={name} width={300}> */}
-            {name}
-            {/* </EllipsisText> */}
+            <EllipsisText widthTooltip={300}>{name}</EllipsisText>
           </Styles.ProductOwnerTitleH2>
         </Styles.ProductOwnerName>
       </Styles.ProductOwner>
@@ -54,9 +51,9 @@ const ProductOwnerData = ({
             )}
 
             <Styles.PostReqDataText color="black" style={{ width: '200px' }}>
-              {/* <EllipsisText id={city} width={300}> */}
-              {city}
-              {/* </EllipsisText> */}
+              <EllipsisText justifyContent={'start'} widthTooltip={300}>
+                {city}
+              </EllipsisText>
             </Styles.PostReqDataText>
 
             {phone && (
